+++ src/seqno_iter.rs
@@ -1,9 +1,8 @@
-use crate::flatfile::FlatFile;
+use crate::{flatfile::FlatFile, SharedMmap};
 use std::{mem::size_of, sync::Arc};
 
 /// This structure allows to iterate over records in the order they were added
-/// to this database. Note that this structure provides only the `Iterator`-like
-/// interface, because `Iterator` items cannot have associated lifetimes.
+/// to this database.
 pub struct SeqNoIter {
     data: Arc<FlatFile>,
     offset: usize,
@@ -14,9 +13,17 @@ impl SeqNoIter {
         Self { data, offset }
     }
 
-    pub fn next(&mut self) -> Option<&[u8]> {
+    fn next_impl(&mut self) -> Option<SharedMmap> {
         let item = self.data.get_record_at_offset(self.offset)?;
         self.offset += item.len() + size_of::<u64>();
         Some(item)
+    }
+}
+
+impl Iterator for SeqNoIter {
+    type Item = SharedMmap;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.next_impl()
     }
 }