+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Public methods now use `SharedMmap` instead of `&[u8]`.
+- `SeqNoIter` now also uses `SharedMmap` which allows it to use the `Iterator`
+  trait.
 
 ## [0.2.0] - 2020-08-04
 ### Removed